refactor: move `NullifierTree` to separated file, renames ans small fixes
