+++ store/src/lib.rs
@@ -2,6 +2,7 @@ pub mod config;
 pub mod db;
 pub mod errors;
 pub mod genesis;
+mod nullifier_tree;
 pub mod server;
 pub mod state;
 pub mod types;
