+++ store/src/state.rs
@@ -9,10 +9,10 @@ use miden_node_utils::formatting::{format_account_id, format_array};
 use miden_objects::{
     crypto::{
         hash::rpo::RpoDigest,
-        merkle::{LeafIndex, Mmr, MmrDelta, MmrPeaks, SimpleSmt, Smt, SmtProof, ValuePath},
+        merkle::{LeafIndex, Mmr, MmrDelta, MmrPeaks, SimpleSmt, SmtProof, ValuePath},
     },
     notes::{NoteMetadata, Nullifier, NOTE_LEAF_DEPTH},
-    AccountError, BlockHeader, Felt, FieldElement, Word, ACCOUNT_TREE_DEPTH, EMPTY_WORD,
+    AccountError, BlockHeader, Word, ACCOUNT_TREE_DEPTH,
 };
 use tokio::{
     sync::{oneshot, Mutex, RwLock},
@@ -23,9 +23,10 @@ use tracing::{error, info, info_span, instrument};
 use crate::{
     db::{AccountInfo, Db, Note, NoteCreated, NullifierInfo, StateSyncUpdate},
     errors::{
-        ApplyBlockError, DatabaseError, GetBlockInputsError, NullifierTreeError,
-        StateInitializationError, StateSyncError,
+        ApplyBlockError, DatabaseError, GetBlockInputsError, StateInitializationError,
+        StateSyncError,
     },
+    nullifier_tree::NullifierTree,
     types::{AccountId, BlockNumber},
     COMPONENT,
 };
@@ -37,82 +38,6 @@ use crate::{
 pub struct TransactionInputs {
     pub account_hash: RpoDigest,
     pub nullifiers: Vec<NullifierInfo>,
-}
-
-/// Nullifier SMT.
-#[derive(Debug, Clone)]
-struct NullifierTree(Smt);
-
-impl NullifierTree {
-    /// Construct new nullifier tree from list of items.
-    pub fn with_entries(
-        entries: impl IntoIterator<Item = (Nullifier, BlockNumber)>
-    ) -> Result<Self, NullifierTreeError> {
-        let leaves = entries.into_iter().map(|(nullifier, block_num)| {
-            (nullifier.inner(), Self::block_num_to_nullifier_value(block_num))
-        });
-
-        let inner = Smt::with_entries(leaves)?;
-
-        Ok(Self(inner))
-    }
-
-    /// Get SMT root.
-    pub fn root(&self) -> RpoDigest {
-        self.0.root()
-    }
-
-    /// Returns an opening of the leaf associated with the given nullifier.
-    pub fn open(
-        &self,
-        nullifier: &Nullifier,
-    ) -> SmtProof {
-        self.0.open(&nullifier.inner())
-    }
-
-    /// Inserts block number in which nullifier was consumed.
-    pub fn insert(
-        &mut self,
-        nullifier: &Nullifier,
-        block_num: BlockNumber,
-    ) -> Result<(), NullifierTreeError> {
-        let value = self.0.insert(nullifier.inner(), Self::block_num_to_nullifier_value(block_num));
-        if value == EMPTY_WORD {
-            return Ok(());
-        }
-
-        Err(NullifierTreeError::NullifierAlreadyExists {
-            nullifier: *nullifier,
-            block_num: Self::nullifier_value_to_block_num(value),
-        })
-    }
-
-    /// Returns block number stored for the given nullifier or `None` if the nullifier wasn't
-    /// consumed.
-    pub fn get_block_num(
-        &self,
-        nullifier: &Nullifier,
-    ) -> Option<BlockNumber> {
-        let value = self.0.get_value(&nullifier.inner());
-        if value == EMPTY_WORD {
-            return None;
-        }
-
-        Some(Self::nullifier_value_to_block_num(value))
-    }
-
-    /// Returns the nullifier's leaf value in the SMT by its block number.
-    fn block_num_to_nullifier_value(block: BlockNumber) -> Word {
-        [Felt::from(block), Felt::ZERO, Felt::ZERO, Felt::ZERO]
-    }
-
-    /// Given the leaf value of the nullifier SMT, returns the nullifier's block number.
-    ///
-    /// There are no nullifiers in the genesis block. The value zero is instead used to signal
-    /// absence of a value.
-    fn nullifier_value_to_block_num(value: Word) -> BlockNumber {
-        value[0].as_int().try_into().expect("invalid block number found in store")
-    }
 }
 
 /// Container for state that needs to be updated atomically.
@@ -577,7 +502,7 @@ async fn load_nullifier_tree(db: &mut Db) -> Result<NullifierTree, StateInitiali
 
     let now = Instant::now();
     let nullifier_tree = NullifierTree::with_entries(nullifiers)
-        .map_err(StateInitializationError::FailedToCreateNullifiersTree)?;
+        .map_err(StateInitializationError::FailedToCreateNullifierTree)?;
     let elapsed = now.elapsed().as_secs();
 
     info!(
@@ -610,28 +535,4 @@ async fn load_accounts(
 
     SimpleSmt::with_leaves(account_data)
         .map_err(StateInitializationError::FailedToCreateAccountsTree)
-}
-
-#[cfg(test)]
-mod tests {
-    use miden_objects::{Felt, ZERO};
-
-    use crate::state::NullifierTree;
-
-    #[test]
-    fn test_nullifier_data_encoding() {
-        let block_num = 123;
-        let nullifier_value = NullifierTree::block_num_to_nullifier_value(block_num);
-
-        assert_eq!(nullifier_value, [Felt::from(block_num), ZERO, ZERO, ZERO])
-    }
-
-    #[test]
-    fn test_nullifier_data_decoding() {
-        let block_num = 123;
-        let nullifier_value = [Felt::from(block_num), ZERO, ZERO, ZERO];
-        let decoded_block_num = NullifierTree::nullifier_value_to_block_num(nullifier_value);
-
-        assert_eq!(decoded_block_num, block_num);
-    }
 }
